requirements.txt
# 智能学习助手 - Python依赖
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
pydantic==2.5.1
tiktoken==0.5.2
openai==1.93.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.2.2
python-dotenv==1.0.0
redis==5.0.1
aioredis==2.0.1 